requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==2.4.0
djangorestframework==3.15.1
docopt==0.6.2
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.19.1
idna==3.7
inflection==0.5.1
ipython==8.12.3
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.18.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9